+++ plugins/widget/plugin.js
@@ -3461,6 +3461,39 @@
 		}
 	} );
 
+	function insertLine( widget, position ) {
+		var elementTag = decodeEnterMode( widget.editor.config.enterMode ),
+			newElement = new CKEDITOR.dom.element( elementTag );
+
+		newElement.appendBogus();
+
+		if ( position === 'after' ) {
+			newElement.insertAfter( widget.wrapper );
+		} else {
+			newElement.insertBefore( widget.wrapper );
+		}
+
+		select( newElement );
+
+		function decodeEnterMode( option ) {
+			if ( option == CKEDITOR.ENTER_BR ) {
+				return 'br';
+			} else if ( option == CKEDITOR.ENTER_DIV ) {
+				return 'div';
+			}
+
+			// Default option - CKEDITOR.ENTER_P.
+			return 'p';
+		}
+
+		function select( element ) {
+			var newRange = widget.editor.createRange();
+
+			newRange.setStart( element, 0 );
+			widget.editor.getSelection().selectRanges( [ newRange ] );
+		}
+	}
+
 	function copyWidgets( editor, isCut ) {
 		var focused = editor.widgets.focused,
 			isWholeSelection,
@@ -3574,6 +3607,9 @@
 	}
 
 	function setupWidget( widget, widgetDef ) {
+		var keystrokeInsertLineBefore = widget.editor.config.widget_keystrokeInsertLineBefore,
+			keystrokeInsertLineAfter = widget.editor.config.widget_keystrokeInsertLineAfter;
+
 		setupWrapper( widget );
 		setupParts( widget );
 		setupEditables( widget );
@@ -3602,16 +3638,28 @@
 		widget.on( 'key', function( evt ) {
 			var keyCode = evt.data.keyCode;
 
+			// Insert a new paragraph before the widget (#4467).
+			if ( keyCode == keystrokeInsertLineBefore ) {
+				insertLine( widget, 'before' );
+				widget.editor.fire( 'saveSnapshot' );
+			}
+			// Insert a new paragraph after the widget (#4467).
+			else if ( keyCode == keystrokeInsertLineAfter ) {
+				insertLine( widget, 'after' );
+				widget.editor.fire( 'saveSnapshot' );
+			}
 			// ENTER.
-			if ( keyCode == 13 ) {
+			else if ( keyCode == 13 ) {
 				widget.edit();
-				// CTRL+C or CTRL+X.
-			} else if ( keyCode == CKEDITOR.CTRL + 67 || keyCode == CKEDITOR.CTRL + 88 ) {
+			}
+			// CTRL+C or CTRL+X.
+			else if ( keyCode == CKEDITOR.CTRL + 67 || keyCode == CKEDITOR.CTRL + 88 ) {
 				copyWidgets( widget.editor, keyCode == CKEDITOR.CTRL + 88 );
 				return; // Do not preventDefault.
-				// Pass chosen keystrokes to other plugins or default fake sel handlers.
-				// Pass all CTRL/ALT keystrokes.
-			} else if ( keyCode in keystrokesNotBlockedByWidget ||
+			}
+			// Pass all CTRL/ALT keystrokes.
+			// Pass chosen keystrokes to other plugins or default fake sel handlers.
+			else if ( keyCode in keystrokesNotBlockedByWidget ||
 				( CKEDITOR.CTRL & keyCode ) ||
 				( CKEDITOR.ALT & keyCode ) ) {
 				return;
@@ -4850,3 +4898,27 @@
  *
  * @property {String} pathName
  */
+
+/**
+ * Defines the keyboard shortcut for inserting a line before selected widget. Default combination
+ * is `Shift+Alt+Enter`. New element tag is based on {@link CKEDITOR.config#enterMode} option.
+ *
+ *		config.widget_keystrokeInsertLineBefore = 'CKEDITOR.SHIFT + 38'; // Shift + Arrow Up
+ *
+ * @since 4.17.0
+ * @cfg {Number} [widget_keystrokeInsertLineBefore=CKEDITOR.SHIFT+CKEDITOR.ALT+13]
+ * @member CKEDITOR.config
+ */
+CKEDITOR.config.widget_keystrokeInsertLineBefore = CKEDITOR.SHIFT + CKEDITOR.ALT + 13;
+
+/**
+ * Defines the keyboard shortcut for inserting a line after selected widget. Default combination
+ * is `Shift+Enter`. New element tag is based on {@link CKEDITOR.config#enterMode} option.
+ *
+ *		config.widget_keystrokeInsertLineAfter = 'CKEDITOR.SHIFT + 40'; // Shift + Arrow Down
+ *
+ * @since 4.17.0
+ * @cfg {Number} [widget_keystrokeInsertLineAfter=CKEDITOR.SHIFT+13]
+ * @member CKEDITOR.config
+ */
+CKEDITOR.config.widget_keystrokeInsertLineAfter = CKEDITOR.SHIFT + 13;